// src/app/page/master/map-produk-objekpajak/map-produk-objekpajak.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapProdukObjekpajak {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}